Improved configuration, compilation and linking process.  Added options to control whether or not to build junctools, and whether to dynamically link in boost.  Improved the python detection.

Added column in tab output to handle sample number

Made GTF loading more flexible in junctools

Fixed a bug interpreting mapsplice junction files

Fixed a bug handling block sizes in bed files

Fixed several travis issues

+++ lib/src/markov_model.cc
@@ -28,12 +28,12 @@ using portcullis::SeqUtils;
 #include <portcullis/ml/markov_model.hpp>
 using portcullis::ml::KMMU;
 
-void portcullis::ml::KmerMarkovModel::train(const vector<string>& input, const uint32_t _order) {
+void portcullis::ml::KmerMarkovModel::train(const vector<string>& input, const uint16_t _order) {
 	order = _order;
 	KMMU temp;
 	for (auto & seq : input) {
 		string s = SeqUtils::makeClean(seq);
-		if (s.size() > order + 1) {
+		if ((uint16_t)s.size() > order + 1) {
 			for (size_t i = order; i < s.size(); i++) {
 				temp[s.substr(i - order, order)][s.substr(i, 1)]++;
 			}
@@ -76,7 +76,7 @@ double portcullis::ml::KmerMarkovModel::getScore(const string& seq) {
 	return log(score);
 }
 
-void portcullis::ml::PosMarkovModel::train(const vector<string>& input, const uint32_t _order) {
+void portcullis::ml::PosMarkovModel::train(const vector<string>& input, const uint16_t _order) {
 	order = _order;
 	PMMU temp;
 	for (auto & seq : input) {
